requirements.txt
langchain
langchain-openai
python-dotenv
rich
pydantic